fix broken tests
